Rename src/core/common to src/core/openjph
OpenJPH headers are included in application code via `#include
<openjph/ojph_version.h>`. The headers are expected to be in a folder
named "openjph". The cmake configuration places them there in the
installation step.

However, if OpenJPH is incorporated into an application via cmake's
`add_subdirectory`, there is no installation step, so there is no
"openjph" folder, leading the `#include <openjph/ojph_version.h>` to
fail.

Renaming the "common" directory to "openjph" resolves the build issue,
since the headers then live inside the source tree in a directory with
same name as the installation. The use of the "common" directory name
is entirely internal to the OpenJPH build, it has no impact on the
installation. The name should be arbitrary, so there should be no
downside to renaming it this way.

Signed-off-by: Cary Phillips <cary@ilm.com>